majorly fixing review code + refactoring

+++ backend/src/main/java/com/popflix/model/Movie.java
@@ -29,7 +29,7 @@ public class Movie {
     private String releaseDate;
     private String title;
     private List<String> video;
-    private Float voteAverage;
+    private Integer voteAverage;
     private Integer voteCount;
     private String imdbId;
     private Long budget;
